+++ src/app/GeneratorTeamsApp.ts
@@ -78,9 +78,24 @@ export class GeneratorTeamsApp extends Generator {
 
     public prompting() {
 
+        interface IAnswers {
+            confirmedAdd: boolean;
+            solutionName: string;
+            whichFolder: string;
+            name: string;
+            developer: string;
+            updateManifestVersion: boolean;
+            manifestVersion: string;
+            mpnId: string;
+            parts: string[];
+            host: string;
+            unitTestsEnabled: boolean;
+            useAzureAppInsights: boolean;
+            azureAppInsightsKey: string;
+        };
         // find out what manifest versions we can use
         const manifestGeneratorFactory = new ManifestGeneratorFactory();
-        const versions: inquirer.objects.ChoiceOption[] = ManifestGeneratorFactory.supportedManifestVersions.filter(version => {
+        const versions: inquirer.objects.ChoiceOption<IAnswers>[] = ManifestGeneratorFactory.supportedManifestVersions.filter(version => {
             // filter out non supprted upgrades
             if (this.options.existingManifest) {
                 const manifestGenerator = manifestGeneratorFactory.createManifestGenerator(version.manifestVersion);
@@ -99,7 +114,7 @@ export class GeneratorTeamsApp extends Generator {
         })
 
         // return the question series
-        return this.prompt(
+        return this.prompt<IAnswers>(
             [
                 {
                     type: 'confirm',
@@ -154,7 +169,7 @@ export class GeneratorTeamsApp extends Generator {
                     type: "confirm",
                     name: "updateManifestVersion",
                     message: `Do you want to change the current manifest version ${this.options.existingManifest && "(" + this.options.existingManifest.manifestVersion + ")"}?`,
-                    when: (answers: any) => this.options.existingManifest && versions.length > 0 && answers.confirmedAdd != false,
+                    when: (answers: IAnswers) => this.options.existingManifest && versions.length > 0 && answers.confirmedAdd != false,
                     default: false
                 },
                 {
@@ -162,15 +177,15 @@ export class GeneratorTeamsApp extends Generator {
                     name: 'manifestVersion',
                     message: 'Which manifest version would you like to use?',
                     choices: versions,
-                    default: versions.find((v: inquirer.objects.ChoiceOption) => v.extra.default) ?
-                        versions.find((v: inquirer.objects.ChoiceOption) => v.extra.default)!.value :
+                    default: versions.find((v: inquirer.objects.ChoiceOption<IAnswers>) => v.extra.default) ?
+                        versions.find((v: inquirer.objects.ChoiceOption<IAnswers>) => v.extra.default)!.value :
                         (versions[0] ? versions[0].value : ""),
-                    when: (answers: any) => (this.options.existingManifest && answers.updateManifestVersion && versions.length > 0) || (!this.options.existingManifest)
+                    when: (answers: IAnswers) => (this.options.existingManifest && answers.updateManifestVersion && versions.length > 0) || (!this.options.existingManifest)
                 },
                 {
                     type: 'input',
                     name: 'mpnId',
-                    message: 'Do you have a Microsoft Partner Id? (Leave blank to skip)',
+                    message: 'Enter your Microsoft Partner Id, if you have one? (Leave blank to skip)',
                     default: undefined,
                     validate: (input: string) => {
                         return input.length <= 10;
@@ -209,7 +224,6 @@ export class GeneratorTeamsApp extends Generator {
                         {
                             name: 'A Message Extension',
                             disabled: () => {
-                                this.log(this.options.existingManifest.composeExtensions[0].commands.length )
                                 if (this.options.existingManifest &&
                                     this.options.existingManifest.composeExtensions &&
                                     this.options.existingManifest.composeExtensions[0] &&
@@ -223,13 +237,13 @@ export class GeneratorTeamsApp extends Generator {
                             value: 'messageextension',
                         }
                     ],
-                    when: (answers: any) => answers.confirmedAdd != false
+                    when: (answers: IAnswers) => answers.confirmedAdd != false
                 },
                 {
                     type: 'input',
                     name: 'host',
                     message: 'The URL where you will host this solution?',
-                    default: (answers: any) => {
+                    default: (answers: IAnswers) => {
                         return `https://${lodash.camelCase(answers.solutionName).toLocaleLowerCase()}.azurewebsites.net`;
                     },
                     validate: Yotilities.validateUrl,
@@ -251,16 +265,16 @@ export class GeneratorTeamsApp extends Generator {
                     type: 'input',
                     name: 'azureAppInsightsKey',
                     message: 'What is the Azure Application Insights Instrumentation Key?',
-                    default: (answers: any) => {
+                    default: (answers: IAnswers) => {
                         return Guid.EMPTY;
                     },
                     validate: (input: string) => {
                         return Guid.isGuid(input);
                     },
-                    when: (answers: any) => answers.useAzureAppInsights,
+                    when: (answers: IAnswers) => answers.useAzureAppInsights,
                 },
             ]
-        ).then((answers: any) => {
+        ).then((answers: IAnswers) => {
             if (answers.confirmedAdd == false) {
                 process.exit(0)
             }