+++ src/app/templates/gulpfile.js
@@ -218,18 +218,21 @@ gulp.task('start-ngrok', (cb) => {
         authtoken: process.env.NGROK_AUTH
     };
 
-    ngrok.connect(conf).then( (url) => {
-        log('[NGROK] Url: ' + url);
 
+    ngrok.connect(conf).then((url) => {
+        log('[NGROK] Url: ' + url);
+        if (!conf.authtoken) {
+            log("[NGROK] You have been assigned a random ngrok URL that will only be available for this session. You wil need to re-upload the Teams manifest next time you run this command.");
+        }
         let hostName = url.replace('http://', '');
         hostName = hostName.replace('https://', '');
-    
+
         log('[NGROK] HOSTNAME: ' + hostName);
         process.env.HOSTNAME = hostName
-    
+
         cb();
-    
-    }).catch( (err) => {
+
+    }).catch((err) => {
         log.error(`[NGROK] Error: ${JSON.stringify(err)}`);
         cb(err.msg);
     });